Finished restructuring source files. It should now be possible to build the core simulator without any dependencies
