+++ src/benchmark/CMakeLists.txt
@@ -1,8 +1,8 @@
 if (BCSIM_ENABLE_CUDA)
     cuda_add_executable(gpu_render_spline_comparison
                         gpu_render_spline_comparison.cu
-                        ../algorithm/cuda_helpers.h
-                        ../bspline.hpp
+                        ../core/algorithm/cuda_helpers.h
+                        ../core/bspline.hpp
                         )
     target_link_libraries(gpu_render_spline_comparison
                           ${CUDA_LIBRARIES}
@@ -12,7 +12,7 @@ if (BCSIM_ENABLE_CUDA)
     
     cuda_add_executable(gpu_memcpy_speedtest
                         gpu_memcpy_speedtest.cu
-                        ../algorithm/cuda_helpers.h
+                        ../core/algorithm/cuda_helpers.h
                         )
     target_link_libraries(gpu_memcpy_speedtest
                           ${CUDA_LIBRARIES}
@@ -21,7 +21,7 @@ if (BCSIM_ENABLE_CUDA)
                           
     cuda_add_executable(gpu_texture_example
                         gpu_texture_example.cu
-                        ../algorithm/cuda_helpers.h
+                        ../core/algorithm/cuda_helpers.h
                         )
     target_link_libraries(gpu_texture_example
                           ${CUDA_LIBRARIES}
@@ -30,11 +30,10 @@ if (BCSIM_ENABLE_CUDA)
                           
     cuda_add_executable(gpu_3d_texture_example
                         gpu_3d_texture_example.cu
-                        ../algorithm/cuda_helpers.h
+                        ../core/algorithm/cuda_helpers.h
                         )
     target_link_libraries(gpu_3d_texture_example
                           ${CUDA_LIBRARIES}
                          )
     install(TARGETS gpu_3d_texture_example DESTINATION bin)
-    
 endif()
