+++ src/core/algorithm/CpuSplineAlgorithm.hpp
@@ -30,7 +30,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #pragma once
 #include <vector>
 #include "CpuBaseAlgorithm.hpp"
-#include "BCSimConfig.hpp"
+#include "../BCSimConfig.hpp"
 
 namespace bcsim {
