+++ src/core/algorithm/GpuSplineAlgorithm2.cuh
@@ -28,7 +28,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
 #pragma once
-#include "LibBCSim.hpp"
+#include "../LibBCSim.hpp"
 #include "GpuBaseAlgorithm.cuh"
 #include "cuda_helpers.h"
 #include "cufft_helpers.h"
