+++ src/qt5gui/MainWindow.hpp
@@ -33,9 +33,9 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <QLabel>
 #include <QSlider>
 #include <QDebug>
-#include "../LibBCSim.hpp"
+#include "../core/LibBCSim.hpp"
 #include "SimTimeManager.hpp"
-#include "ScanGeometry.hpp"
+#include "../utils/ScanGeometry.hpp"
 
 // Forward decl.
 class GLVisualizationWidget;
