+++ src/qt5gui/SimulationParamsWidget.hpp
@@ -29,7 +29,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #pragma once
 #include <QWidget>
-#include "BeamProfile.hpp"
+#include "../core/BeamProfile.hpp"
 
 // Forward decl.
 class QDoubleSpinBox;
